requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pdfplumber==0.11.4
PyMuPDF==1.24.14
openai==1.57.2
python-multipart==0.0.19
pydantic==2.10.3
numpy==2.2.0
python-dotenv==1.0.1
rich==13.9.4
click==8.1.8
pytest==8.3.4
httpx==0.28.1
